WIP factor out test_helpers phase1
